feat: 팀장은 팀에서 퇴출할 수 없는 규칙 작성

+++ tlog-was/src/main/java/com/se/Tlog/domain/Team/domain/Team.java
@@ -76,6 +76,8 @@ public class Team {
 			throw new CustomException(ErrorType.TEAM_USER_NOT_FOUND);
 		if (repoService.countMemberInTeam(id) == 1)
 			throw new CustomException(ErrorType.TEAM_CANNOT_BE_ORPHAN);
+		if (repoService.isLeader(id, user.getId()))
+		    throw new CustomException(ErrorType.CANNOT_REMOVE_TEAM_LEADER);
 		
 		repoService.deleteUserInTeam(this.id, user.getId());
 		// 기타 팀원 삭제 후 처리내용
